feat(backend-modules): refactor, add suppression checkbox, allow for only certain schema types
