+++ packages/backend-modules/documents/graphql/resolvers/Document.js
@@ -21,10 +21,11 @@ const {
 const {
   extractIdsFromNode,
   loadLinkedMetaData,
-  getParsedDocumentId,
 } = require('@orbiting/backend-modules-search/lib/Documents')
 
-const { ASSETS_SERVER_BASE_URL } = process.env
+const {
+  processMeta: processSyntheticReadAloudInMeta,
+} = require('@orbiting/backend-modules-publikator/lib/Derivative/SyntheticReadAloud')
 
 const addTeaserContentHash = (nodes) => {
   nodes.forEach((node) => {
@@ -100,46 +101,7 @@ module.exports = {
       await processRepoImageUrlsInMeta(doc.content, addFormatAuto)
     }
 
-    // @TODO: Move PoC to /lib
-    const { audioSource } = meta
-    if (!audioSource) {
-      const { repoId, commitId, versionName } = getParsedDocumentId(doc.id)
-      if (versionName !== 'preview') {
-        const derivatives = await context.loaders.Derivative.byCommitId.load(
-          commitId,
-        )
-        const synthesizedAudio = derivatives.find(
-          (d) => d.type === 'SyntheticReadAloud' && d.status === 'Ready',
-        )
-        if (synthesizedAudio) {
-          const { result } = synthesizedAudio
-          if (!result) {
-            return meta
-          }
-
-          const { audioDuration, s3 } = result
-          if (!audioDuration || !s3) {
-            return meta
-          }
-
-          const { key, bucket } = s3
-          if (!key || !bucket) {
-            return meta
-          }
-
-          meta.audioSource = {
-            mediaId: Buffer.from(
-              [repoId, 'SyntheticReadAloud'].join('/'),
-            ).toString('base64'),
-            kind: 'syntheticReadAloud',
-            mp3: `${ASSETS_SERVER_BASE_URL}/s3/${bucket}/${key}`,
-            durationMs: Math.round(1000 * audioDuration),
-          }
-        }
-      }
-    }
-
-    return meta
+    return processSyntheticReadAloudInMeta(meta, doc, context)
   },
   async children(
     doc,
