+++ packages/backend-modules/publikator/.gitignore
@@ -1,4 +1,5 @@
 lib/cache/**/*.js
+lib/Derivative/**/*.js
 lib/phases.js
 loaders/**/*.js
 script/*.js
