+++ packages/backend-modules/publikator/graphql/resolvers/_mutations/generateDerivative.js
@@ -2,7 +2,10 @@ const {
   Roles: { ensureUserHasRole },
 } = require('@orbiting/backend-modules-auth')
 
-const { handleSyntheticReadAloud } = require('../../../lib/Document')
+const {
+  canDerive: canDeriveSyntheticReadAloud,
+  derive: deriveSyntheticReadAloud,
+} = require('../../../lib/Derivative/SyntheticReadAloud')
 
 const { document: getDocument } = require('../Commit')
 
@@ -11,31 +14,24 @@ module.exports = async (_, { commitId }, context) => {
   ensureUserHasRole(user, 'editor')
 
   const commit = await loaders.Commit.byId.load(commitId)
-  // const { meta: repoMeta } = await context.loaders.Repo.byId.load(repoId)
+  if (!canDeriveSyntheticReadAloud(commit.meta.template)) {
+    throw new Error(t('api/publikator/generateDerivative/canNot'))
+  }
 
-  // load and check document
-  const doc = await getDocument(
-    commit, // { id: commitId, repo: { id: repoId }, repoId },
-    {},
-    context,
-  )
-
-  const derivative = await handleSyntheticReadAloud(doc, context)
-
+  const doc = await getDocument(commit, {}, context)
+  const derivative = await deriveSyntheticReadAloud(doc, context)
   if (!derivative) {
     throw new Error(t('api/publikator/generateDerivative/unable'))
   }
 
-  if (derivative) {
-    await pubsub.publish('repoChange', {
-      repoChange: {
-        repoId: commit.repoId,
-        mutation: 'UPDATED',
-        commit,
-        derivative,
-      },
-    })
-  }
+  await pubsub.publish('repoChange', {
+    repoChange: {
+      repoId: commit.repoId,
+      mutation: 'UPDATED',
+      commit,
+      derivative,
+    },
+  })
 
   return derivative
 }
